refactoring code

+++ app/halaman/pengguna/admin/index.php
@@ -13,7 +13,7 @@
         <section class="section">
             <div class="card">
                 <div class="card-body">
-                    <table class="table table-striped" id="table1">
+                    <table class="table table-striped">
                         <thead>
                             <tr>
                                 <th class="no-td">No</th>
@@ -21,22 +21,28 @@
                                 <th class="text-center no-td"></th>
                             </tr>
                         </thead>
+                        <?php
+                        $data = $mysqli->query('SELECT * FROM pengguna WHERE status="ADMIN" ORDER BY username');
+                        $no = 1;
+                        ?>
                         <tbody>
-                            <?php
-                            $data = $mysqli->query('SELECT * FROM pengguna WHERE status="ADMIN" ORDER BY username');
-                            $no = 1;
-                            ?>
-                            <?php while ($row = $data->fetch_assoc()) : ?>
+                            <?php if ($data->num_rows) : ?>
+                                <?php while ($row = $data->fetch_assoc()) : ?>
+                                    <tr>
+                                        <td class="text-center"><?= $no++; ?></td>
+                                        <td class="text-center"><?= $row['username']; ?></td>
+                                        <td class="no-td">
+                                            <a href="?halaman=edit_admin&id=<?= $row['id']; ?>" class="btn btn-warning btn-sm text-white"><i class="fas fa-edit"></i></a>
+                                            <a id="tombol-hapus" href="?halaman=hapus_admin&id=<?= $row['id']; ?>" class="btn btn-danger btn-sm" data-text="Akun admin hanya dapat ditambahkan oleh admin lainnya!" data-button-text="Hapus Admin!">
+                                                <i class="fas fa-trash-alt"></i></a>
+                                        </td>
+                                    </tr>
+                                <?php endwhile; ?>
+                            <?php else : ?>
                                 <tr>
-                                    <td class="text-center"><?= $no++; ?></td>
-                                    <td class="text-center"><?= $row['username']; ?></td>
-                                    <td class="no-td">
-                                        <a href="?halaman=edit_admin&id=<?= $row['id']; ?>" class="btn btn-warning btn-sm text-white"><i class="fas fa-edit"></i></a>
-                                        <a id="tombol-hapus" href="?halaman=hapus_admin&id=<?= $row['id']; ?>" class="btn btn-danger btn-sm" data-text="Akun admin hanya dapat ditambahkan oleh admin lainnya!" data-button-text="Hapus Admin!">
-                                            <i class="fas fa-trash-alt"></i></a>
-                                    </td>
+                                    <td class="text-center" colspan="4">Data Tidak Ada</td>
                                 </tr>
-                            <?php endwhile; ?>
+                            <?php endif; ?>
                         </tbody>
                     </table>
                 </div>
